Simplify Box demo structure
